Create @automattic/sentry package
