+++ client/boot/common.js
@@ -2,6 +2,7 @@ import accessibleFocus from '@automattic/accessible-focus';
 import config from '@automattic/calypso-config';
 import { getUrlParts } from '@automattic/calypso-url';
 import { getLanguageSlugs } from '@automattic/i18n-utils';
+import { addBreadcrumb } from '@automattic/sentry';
 import debugFactory from 'debug';
 import page from 'page';
 import ReactDom from 'react-dom';
@@ -24,7 +25,7 @@ import { checkFormHandler } from 'calypso/lib/protect-form';
 import { setReduxStore as setReduxBridgeReduxStore } from 'calypso/lib/redux-bridge';
 import { getSiteFragment, normalize } from 'calypso/lib/route';
 import { isLegacyRoute } from 'calypso/lib/route/legacy-routes';
-import { addBreadcrumb, initSentry } from 'calypso/lib/sentry';
+import { initSentry } from 'calypso/lib/sentry';
 import { hasTouch } from 'calypso/lib/touch-detect';
 import { isOutsideCalypso } from 'calypso/lib/url';
 import { JETPACK_PRICING_PAGE } from 'calypso/lib/url/support';
